docs(example:file-feed): - reformat to tutorial structure - fix linting - update npm scripts - remove duplicated code (parts of it, there is still more) - use signalling server through DNS - clean up - uncheck checked dist version of js-ipfs - fix: remove ipfs-daemon out of the equation - fix: use spawnNode
