+++ examples/file-feed/complete/package.json
@@ -1,15 +1,14 @@
 {
-  "name": "ipfe",
-  "version": "0.1.0",
+  "name": "file-feed",
+  "version": "0.0.0",
+  "bin": "src/ff-cli/bin.js",
   "devDependencies": {
     "buffer": "^5.0.2",
     "react-scripts": "0.8.5",
     "uglify-js": "github:mishoo/UglifyJS2#harmony"
   },
   "dependencies": {
-    "browserify-zlib-next": "^1.0.0",
-    "ipfs": "^0.21.0",
-    "ipfs-daemon": "^0.3.0-beta.18",
+    "browserify-zlib-next": "^1.0.1",
     "mime": "^1.3.4",
     "orbit-db": "^0.16.1",
     "react": "^15.4.2",
@@ -18,7 +17,6 @@
     "react-router": "^3.0.1"
   },
   "scripts": {
-    "postinstall": "cp config/webpack.config.dev.js node_modules/react-scripts/config/webpack.config.dev.js",
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test --env=jsdom",
