+++ examples/file-feed/complete/src/browser-app/Feed.js
@@ -4,30 +4,30 @@ import { toFileInfo } from './utils'
 import './Feed.css'
 
 class Feed extends Component {
-  render() {
+  render () {
     const { name, files, peers, onOpenFile } = this.props
 
-    const filesList = files.length > 0 
+    const filesList = files.length > 0
       ? files.map((e) => {
-          const fileInfo = toFileInfo(e)
-          return <File key={fileInfo.hash + fileInfo.ts}
-                       fileInfo={fileInfo}
-                       onOpenFile={onOpenFile}/>
-        })
+        const fileInfo = toFileInfo(e)
+        return <File key={fileInfo.hash + fileInfo.ts}
+          fileInfo={fileInfo}
+          onOpenFile={onOpenFile} />
+      })
       : <div className='Feed-empty'>
           This feed is empty.
         </div>
 
     const peerCount = peers ? peers.length : 0
     const peerCounter = peers && peerCount > 0
-      ? <div className='Feed-peers' 
-             onClick={this.props.onShowPeers}>
-          {peerCount} {peerCount === 1 ? 'peer' : 'peers'}
-        </div>
+      ? <div className='Feed-peers'
+        onClick={this.props.onShowPeers}>
+        {peerCount} {peerCount === 1 ? 'peer' : 'peers'}
+      </div>
       : <div className='Feed-peers'
-             onClick={this.props.onShowPeers}>
-          <i>Searching for peers...</i>
-        </div>
+        onClick={this.props.onShowPeers}>
+        <i>Searching for peers...</i>
+      </div>
 
     return (
       <div className='Feed'>
