+++ examples/file-feed/complete/src/browser-app/Preview.js
@@ -8,7 +8,7 @@ let dataStore
 class Preview extends Component {
   constructor (props) {
     super(props)
-    this.state = { 
+    this.state = {
       data: null,
       progress: 'Loading...'
     }
@@ -38,10 +38,10 @@ class Preview extends Component {
     dataStore.removeListener('load', this.handleLoadProgress)
   }
 
-  handleLoadProgress (hash, bytes){
+  handleLoadProgress (hash, bytes) {
     const percentage = Math.ceil(bytes / this.props.size * 100)
-    const progress = this.props.size > 10000 || percentage === 100 
-      ? 'Loading ' + percentage + '%' 
+    const progress = this.props.size > 10000 || percentage === 100
+      ? 'Loading ' + percentage + '%'
       : this.state.progress
 
     this.setState({ progress: progress })
@@ -60,35 +60,35 @@ class Preview extends Component {
 
     if (type === 'image') {
       output = <img style={{ maxWidth: '90%', maxHeight: '80%' }}
-                    src={src} 
-                    alt=''/>
+        src={src}
+        alt='' />
     } else if (type === 'audio') {
-      output = <audio style={{ width: '60%', maxHeight: '80%' }} 
-                      src={src} 
-                      controls 
-                      autoPlay={true} 
-                      onClick={stopEventBubbling}/>
+      output = <audio style={{ width: '60%', maxHeight: '80%' }}
+        src={src}
+        controls
+        autoPlay
+        onClick={stopEventBubbling} />
     } else if (type === 'video') {
-      output = <video style={{ maxWidth: '90%', maxHeight: '80%' }} 
-                      src={src} 
-                      controls 
-                      autoPlay={true} 
-                      onClick={stopEventBubbling}/>
+      output = <video style={{ maxWidth: '90%', maxHeight: '80%' }}
+        src={src}
+        controls
+        autoPlay
+        onClick={stopEventBubbling} />
     } else {
       output = this.state.data
-        ? <pre className="Preview-content" style={{ width: '80%' }}>{src}</pre>
+        ? <pre className='Preview-content' style={{ width: '80%' }}>{src}</pre>
         : null
     }
 
-    const statusText = !src 
-      ? <div className="Preview-loading"><br/>{progress}</div>
-      : <div className="Preview-info">
-          <span>{name}</span>
-          <span>{hash}</span>
-        </div>
+    const statusText = !src
+      ? <div className='Preview-loading'><br />{progress}</div>
+      : <div className='Preview-info'>
+        <span>{name}</span>
+        <span>{hash}</span>
+      </div>
 
     return (
-      <div className="Preview" onClick={onClick}>
+      <div className='Preview' onClick={onClick}>
         {statusText}
         {output}
       </div>
