+++ examples/file-feed/complete/src/browser-app/index.js
@@ -6,13 +6,10 @@ import './index.css'
 
 const Index = () => (
   <Router history={browserHistory}>
-    <Route path="/" component={App}/>
-    <Route path="/:hash" component={App}/>
-    <Route path="/open/:file" component={App}/>
+    <Route path='/' component={App} />
+    <Route path='/:hash' component={App} />
+    <Route path='/open/:file' component={App} />
   </Router>
 )
 
-ReactDOM.render(
-  <Index />,
-  document.getElementById('root')
-)
+ReactDOM.render(<Index />, document.getElementById('root'))
