+++ examples/file-feed/complete/src/index.js
@@ -1,0 +1,3 @@
+'use strict'
+
+require('./browser-app')
